modularize VinylCircle, GenrePieChart
