+++ src/app/user/[slug]/UserContent.jsx
@@ -4,121 +4,9 @@ Contains info about a user's top genres, top artists, top songs, and song analys
 */
 
 import { ResponsiveRadar } from "@nivo/radar";
-import { ResponsivePie } from "@nivo/pie";
-import { useState, useEffect, useRef } from "react";
-import "@/app/globals.css";
 import PropTypes from "prop-types";
-
-function VinylCircle({ centerCircleColor, width }) {
-  const newWidth = Math.min((width - 280) / 2, 160);
-  const radii = [];
-  if (newWidth > 0) {
-    for (let i = newWidth - 1; i > 10; i -= 3) {
-      radii.push(i);
-    }
-  }
-
-  return (
-    <svg width={Math.min(400, width)} height="400" data-testid="VinylCircle">
-      {radii.map((radius) => (
-        <circle
-          key={radius}
-          r={radius}
-          cx="200"
-          cy="200"
-          fill="none"
-          stroke="black"
-          strokeWidth="1.35"
-        />
-      ))}
-      <circle
-        r={Math.round(newWidth * 0.15625)}
-        cx="200"
-        cy="200"
-        fill="none"
-        stroke={centerCircleColor}
-        strokeWidth={Math.round(newWidth * 0.2)}
-      />
-      <circle
-        r={Math.round(newWidth * 0.275)}
-        cx="200"
-        cy="200"
-        fill="none"
-        stroke="black"
-        strokeWidth={Math.round(newWidth * 0.05)}
-      />
-    </svg>
-  );
-}
-
-function GenrePieChart({ data, centerCircleColor }) {
-  const [divWidth, setDivWidth] = useState(0); // Step 1: State for storing div width
-  const divRef = useRef(null); // Step 2: Ref for the div
-
-  // Step 3: Effect hook for setting and updating div width
-  useEffect(() => {
-    // Function to update div width
-    const updateWidth = () => {
-      if (divRef.current) {
-        setDivWidth(divRef.current.offsetWidth); // Update div width
-      }
-    };
-
-    window.addEventListener("resize", updateWidth); // Add resize event listener
-    updateWidth(); // Initial update
-
-    return () => window.removeEventListener("resize", updateWidth); // Cleanup
-  }, []);
-
-  return (
-    <div
-      ref={divRef}
-      style={{ height: 440, position: "relative" }}
-      data-testid="GenrePieChart"
-    >
-      <ResponsivePie
-        data={data}
-        margin={{ top: 70, right: 140, bottom: 50, left: 140 }}
-        innerRadius={0.3}
-        keys={["value"]}
-        colors={["#8D97A9", "#BDC8DA", "#121D2A", "#737D8E", "#1F2937"]}
-        arcLinkLabelsTextColor="#333333"
-        arcLinkLabelsThickness={2}
-        arcLinkLabelsColor={{ from: "color" }}
-        enableArcLabels={false}
-        arcLabelsTextColor={{ from: "color", modifiers: [["darker", 2]] }}
-        isInteractive={false}
-        animate={false}
-        legends={[]}
-        theme={{
-          text: {
-            fontSize: 20,
-            fill: "#333333",
-            outlineWidth: 0,
-            outlineColor: "transparent",
-            fontFamily: "Koulen",
-          },
-        }}
-      />
-      <div
-        style={{
-          position: "absolute",
-          top: 20,
-          right: 0,
-          bottom: 0,
-          left: Math.min(0, divWidth - 400),
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          justifyContent: "center",
-          textAlign: "center",
-        }}
-      >
-        <VinylCircle centerCircleColor={centerCircleColor} width={divWidth} />
-      </div>
-    </div>
-  );
-}
+import "@/app/globals.css";
+import GenrePieChart from "@/shared/GenrePieChart";
 
 function UserContent({ userData, shareCassette }) {
   // Convert object to array of { id: genre, value: frequency } objects
@@ -233,7 +121,7 @@ function UserContent({ userData, shareCassette }) {
                   fill: "#333333",
                   outlineWidth: 10,
                   outlineColor: "transparent",
-                  fontFamily: "Koulen",
+                  fontFamily: "var(--font-koulen)",
                 },
               }}
               colors={"#39466B"}
@@ -288,19 +176,3 @@ UserContent.propTypes = {
   }).isRequired,
   shareCassette: PropTypes.func.isRequired,
 };
-
-VinylCircle.propTypes = {
-  centerCircleColor: PropTypes.string,
-  width: PropTypes.number,
-};
-GenrePieChart.propTypes = {
-  data: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      value: PropTypes.number.isRequired,
-    }),
-  ).isRequired,
-  centerCircleColor: PropTypes.string,
-};
-
-export { VinylCircle, GenrePieChart };
